ada post processing block as a module
Former-commit-id: aa3e8d1ef21fb6118d6e5aafe9319bcc28fda0e2

+++ python/pybirales/modules/channeliser.py
@@ -81,7 +81,7 @@ class PFB(ProcessingModule):
         # Generate output blob
         return ChannelisedBlob(self._config, [('nbeams', nstreams),
                                               ('nchans', self._nchans * input_shape['nsubs']),
-                                              ('nsamp', input_shape['nsamp'] / self._nchans)],
+                                              ('nsamp', int(input_shape['nsamp'] / self._nchans))],
                                datatype=datatype)
 
     def _initialise(self, nsamp, nbeams, nsubs):
